feat: add game module with map and battle init
