+++ README.md
@@ -16,7 +16,8 @@ Protótipo de jogo de tabuleiro com cartas e turnos, escrito em HTML, CSS e Java
 ## Estrutura
 
 - `map.html`: mapa de fases com indicação do estágio atual.
-- `js/map.js`: renderização do mapa e controle do estágio.
+- `js/game.js`: lógica de mapa e inicialização de batalha.
+- `js/map.js`: configuração da tela do mapa.
 - `index.html`: marcação principal do tabuleiro.
 - Diretório `css/`: estilos da interface.
 - `js/main.js`: lógica de jogo e controle de turnos.
